// src/hooks/orval/model/participantsRecordRegistrationStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */

export type ParticipantsRecordRegistrationStatus = typeof ParticipantsRecordRegistrationStatus[keyof typeof ParticipantsRecordRegistrationStatus];


export const ParticipantsRecordRegistrationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;